refactor dialog
