+++ flutter/lib/desktop/pages/connection_page.dart
@@ -19,6 +19,8 @@ import '../../common/widgets/peer_tab_page.dart';
 import '../../models/platform_model.dart';
 import '../widgets/button.dart';
 
+import 'package:flutter_hbb/common/widgets/dialog.dart';
+
 /// Connection page for connecting to a remote peer.
 class ConnectionPage extends StatefulWidget {
   const ConnectionPage({Key? key}) : super(key: key);
