+++ src/lang/de.rs
@@ -480,5 +480,8 @@ pub static ref T: std::collections::HashMap<&'static str, &'static str> =
         ("identical_file_tip", "Diese Datei ist identisch mit der Datei der Gegenstelle."),
         ("show_monitors_tip", "Monitore in der Symbolleiste anzeigen"),
         ("View Mode", "Ansichtsmodus"),
+        ("Enter RustDesk password", ""),
+        ("Remember RustDesk password", ""),
+        ("Login Required", ""),
     ].iter().cloned().collect();
 }
