+++ src/lang/tw.rs
@@ -480,5 +480,8 @@ pub static ref T: std::collections::HashMap<&'static str, &'static str> =
         ("identical_file_tip", "此檔案與對方的檔案一致"),
         ("show_monitors_tip", "在工具列中顯示顯示器"),
         ("View Mode", "瀏覽模式"),
+        ("Enter RustDesk password", ""),
+        ("Remember RustDesk password", ""),
+        ("Login Required", ""),
     ].iter().cloned().collect();
 }
